serve: Factor out startDebugServer into pkg
We want to use it for other go services.
